requirements.txt
ccxt==4.2.41
python-telegram-bot==13.15
flask==2.3.3
requests==2.31.0
flask_sqlalchemy==3.1.1
sqlalchemy==2.0.23